refactoring and ect
