+++ src/main/java/com/students/demo/MyBean1.java
@@ -7,14 +7,24 @@ import lombok.extern.log4j.Log4j2;
 public class MyBean1 implements MyInterface  {
     @Override
     public void doSomething() {
-        Thread thread = new Thread(() -> {
-            System.out.println("Thread1 bean is initialized");
-            System.out.println("Doing something in MyBean1");
+        /*Thread thread = new Thread(() -> {
+            //log.info("Thread1 bean is initialized");
+            //System.out.println
+            log.info("Doing something in MyBean1");
         });
-        thread.start();
-
-        //System.out.println("Doing something in MyBean1");
+        thread.start();*/
+        //log.info("Doing something in MyBean1");
     }
+    @PostConstruct
+    public void postConstruct() {
+        log.info("MyBean1 bean is initialized");
+    }
+
+    @PreDestroy
+    public void preDestroy() {
+        log.info("MyBean1 bean is destroyed");
+    }
+
 }
