+++ src/main/java/com/students/demo/MyBean2.java
@@ -1,17 +1,29 @@
 package com.students.demo;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
-import org.springframework.stereotype.Component;
 import lombok.extern.log4j.Log4j2;
 @Log4j2
 public class MyBean2 implements MyInterface  {
+
     @Override
     public void doSomething() {
-        Thread thread = new Thread(() -> {
-            System.out.println("Thread2 bean is initialized");
-            System.out.println("Doing something in MyBean2");
+        /*Thread thread = new Thread(() -> {
+            //log.info("Thread1 bean is initialized");
+            //System.out.println
+            log.info("Doing something in MyBean1");
         });
-        thread.start();
+        thread.start();*/
+        //log.info("Doing something in bean" );
     }
+    @PostConstruct
+    public void postConstruct() {
+        log.info("MyBean2 bean is initialized");
+    }
+
+    @PreDestroy
+    public void preDestroy() {
+        log.info("MyBean2 bean is destroyed");
+    }
+
 }
